infra

+++ APPInfraEstructure/RepositoryRead/ConcreteQuerys/Migration/yFileUpload/IyFileUploadReadQuerys.cs
@@ -11,7 +11,7 @@ namespace IQuery.Read
         public QueryModel ExistsByTypeQuery(string value );
         public QueryModel ExistsByStatusQuery(int value );
         public QueryModel ExistsByFilePathQuery(string value );
-        public QueryModel ExistsByFileSizeQuery(int value );
+        public QueryModel ExistsByFileSizeQuery(long value );
         public QueryModel ExistsByContentTypeQuery(string value );
         public QueryModel ExistsByCreatedAtQuery(DateTime value );
         public QueryModel ExistsByCompletedAtQuery(DateTime value );
@@ -24,7 +24,7 @@ namespace IQuery.Read
         public QueryModel FirstByTypeQuery(string value );
         public QueryModel FirstByStatusQuery(int value );
         public QueryModel FirstByFilePathQuery(string value );
-        public QueryModel FirstByFileSizeQuery(int value );
+        public QueryModel FirstByFileSizeQuery(long value );
         public QueryModel FirstByContentTypeQuery(string value );
         public QueryModel FirstByCreatedAtQuery(DateTime value );
         public QueryModel FirstByCompletedAtQuery(DateTime value );
