+++ AppDominio/Dominio/Entitys/Migration/yFileUpload/IyFileUploadEntity.cs
@@ -15,7 +15,7 @@
     string Type { get; set; }
     int Status { get; set; }
     string FilePath { get; set; }
-    int? FileSize { get; set; }
+    long? FileSize { get; set; }
     string ContentType { get; set; }
     DateTime CreatedAt { get; set; }
     DateTime? CompletedAt { get; set; }
