+++ AppDominio/Dominio/Entitys/Migration/yFileUpload/yFileUploadFactory.cs
@@ -9,7 +9,7 @@
                                     public yFileUploadFactory(Dominio.Interfaces.ILogger logger)
                                     {
                                         _logger = logger;
-                                    } public IyFileUploadEntity Create(int? id, string idempotencykey, string type, int status, string filepath, int? filesize, string contenttype, DateTime createdat, DateTime? completedat )
+                                    } public IyFileUploadEntity Create(int? id, string idempotencykey, string type, int status, string filepath, long? filesize, string contenttype, DateTime createdat, DateTime? completedat )
                             {
                             var entity = new yFileUploadEntity(id, idempotencykey, type, status, filepath, filesize, contenttype, createdat, completedat );
